+++ src/main/untangled/easy_server.clj
@@ -2,7 +2,7 @@
   (:require
     [com.stuartsierra.component :as component]
     [clojure.set :as set]
-    [clojure.spec :as s]
+    [clojure.spec.alpha :as s]
     [clojure.java.io :as io]
     [om.next.server :as om]
     [taoensso.timbre :as timbre]