+++ src/test/untangled/client/ui_spec.clj
@@ -1,6 +1,5 @@
 (ns untangled.client.ui-spec
   (:require
-    [clojure.spec :as s]
     [clojure.test :as t]
     [om.next :as om]
     [untangled-spec.core :refer